API: Move mirror out of epics sub-directory
